requirements.txt
aiohttp
python-socketio
tk
PyPDF2